Add tpp-mlir as a submodule
This commit adds tpp-mlir as a submodule and the appropriate CMake flags
to build it inside IREE. The tpp-mlir repository has the cmake logic to
be built inside IREE for now. In the future, we want to move that into
IREE, so that we can build it inside other frameworks/compilers without
having to carry CMake rules for all of them.